+++ src/ArduinoJson/Object/ObjectImpl.hpp
@@ -13,26 +13,26 @@ template <typename TObject>
 template <typename TString>
 inline ArrayRef ObjectShortcuts<TObject>::createNestedArray(
     const TString& key) const {
-  return impl()->getOrCreate(key).template to<ArrayRef>();
+  return impl()->getOrAddMember(key).template to<ArrayRef>();
 }
 
 template <typename TObject>
 template <typename TChar>
 inline ArrayRef ObjectShortcuts<TObject>::createNestedArray(TChar* key) const {
-  return impl()->getOrCreate(key).template to<ArrayRef>();
+  return impl()->getOrAddMember(key).template to<ArrayRef>();
 }
 
 template <typename TObject>
 template <typename TString>
 inline ObjectRef ObjectShortcuts<TObject>::createNestedObject(
     const TString& key) const {
-  return impl()->getOrCreate(key).template to<ObjectRef>();
+  return impl()->getOrAddMember(key).template to<ObjectRef>();
 }
 
 template <typename TObject>
 template <typename TChar>
 inline ObjectRef ObjectShortcuts<TObject>::createNestedObject(
     TChar* key) const {
-  return impl()->getOrCreate(key).template to<ObjectRef>();
+  return impl()->getOrAddMember(key).template to<ObjectRef>();
 }
 }  // namespace ARDUINOJSON_NAMESPACE