+++ src/ArduinoJson/Variant/VariantFunctions.hpp
@@ -147,19 +147,19 @@ inline CollectionData *variantToObject(VariantData *var) {
 }
 
 inline NO_INLINE VariantData *variantAdd(VariantData *var, MemoryPool *pool) {
-  return var != 0 ? var->add(pool) : 0;
+  return var != 0 ? var->addElement(pool) : 0;
 }
 
 template <typename TChar>
 NO_INLINE VariantData *variantGetOrCreate(VariantData *var, TChar *key,
                                           MemoryPool *pool) {
-  return var != 0 ? var->getOrCreate(adaptString(key), pool) : 0;
+  return var != 0 ? var->getOrAddMember(adaptString(key), pool) : 0;
 }
 
 template <typename TString>
 NO_INLINE VariantData *variantGetOrCreate(VariantData *var, const TString &key,
                                           MemoryPool *pool) {
-  return var != 0 ? var->getOrCreate(adaptString(key), pool) : 0;
+  return var != 0 ? var->getOrAddMember(adaptString(key), pool) : 0;
 }
 
 }  // namespace ARDUINOJSON_NAMESPACE