+++ test/JsonVariant/add.cpp
@@ -12,13 +12,6 @@ TEST_CASE("JsonVariant::add()") {
   DynamicJsonDocument doc(4096);
   JsonVariant var = doc.to<JsonVariant>();
 
-  SECTION("No argument") {
-    JsonVariant nested = var.add();
-
-    REQUIRE(var.is<JsonArray>() == true);
-    REQUIRE(nested.isNull() == true);
-  }
-
   SECTION("integer") {
     var.add(42);